+++ sdk/go/core/ydk/types/types.go
@@ -389,100 +389,6 @@ func GetYLeafs(entityData *CommonEntityData) []YLeaf {
 	return leafs
 }
 
-// YList utilities to access list elements by key
-func buildKeyList(entity Entity) []interface{} {
-	keySlice := entity.GetEntityData().YListKeys
-	key_list := []interface{}{}
-	s := reflect.ValueOf(entity).Elem()
-
-	for _, key := range keySlice {
-		keyValueField := s.FieldByName(key)
-		if keyValueField.IsValid() {
-			f := keyValueField.Interface()
-			val := reflect.ValueOf(f)
-			switch val.Kind() {
-			case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-				intVal := val.Int()
-				key_list = append(key_list, intVal)
-			case reflect.String:
-				strVal := val.String()
-				key_list = append(key_list, strVal)
-			default:
-				sVal := fmt.Sprintf("%v", val)
-				key_list = append(key_list, sVal)
-			}
-		}
-	}
-	return key_list
-}
-
-func keysToStr(key_list []interface{}) string {
-	var key_str string
-	for _, key := range key_list {
-		if len(key_str) > 0 {
-			key_str += ","
-		}
-		key_str += fmt.Sprintf("%v", key)
-	}
-	return key_str
-}
-
-func toEntitySlice(slice interface{}) []Entity {
-	s := reflect.ValueOf(slice)
-	if s.Kind() != reflect.Slice {
-		panic("ToEntitySlice() given a non-slice parameter type")
-	}
-	ret := make([]Entity, s.Len())
-	for i := 0; i < s.Len(); i++ {
-		ent, ok := s.Index(i).Interface().(Entity)
-		if !ok {
-			panic("Slice element is not of type Entity. Cannot convert.")
-		}
-		ret[i] = ent
-	}
-	return ret
-}
-
-func GetFromList(iSlice interface{}, keys ... interface{}) Entity {
-	eSlice := toEntitySlice(iSlice)
-	if len(keys) == 0 {
-		return nil
-	}
-	var keyToCmp string
-	if reflect.ValueOf(keys[0]).Kind() == reflect.Slice {
-		keyToCmp = keysToStr(keys[0].([]interface{}))
-	} else {
-		keyToCmp = keysToStr(keys)
-	}
-	for _, elem := range eSlice {
-		keyList := buildKeyList(elem)
-		if len(keyList) > 0 {
-			key := keysToStr(keyList)
-			if key == keyToCmp {
-				return elem
-			}
-		}
-	}
-	return nil
-}
-
-func GetListKeys(iSlice interface{}) []interface{} {
-	eSlice := toEntitySlice(iSlice)
-	keyList := []interface{}{}
-	for _, ient := range eSlice {
-		ent := ient.(Entity)
-		key := buildKeyList(ent)
-		if len(key) == 1 {
-			keyList = append(keyList, key[0])
-			continue
-		}
-		if len(key) > 1 {
-			keyList = append(keyList, key)
-		}
-	}
-	return keyList
-}
-
 // GetSegmentPath returns the given entity's segment path
 func GetSegmentPath(entity Entity) string {
 	return entity.GetEntityData().SegmentPath
